requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
